update: move config to separated folder for better management
