+++ cmd/main.go
@@ -4,116 +4,23 @@ import (
 	"context"
 	"flag"
 	"fmt"
-	"io"
 	"net/http"
 	"os"
 	"os/signal"
-	"strings"
+	"strconv"
 	"syscall"
 	"time"
 
+	"go.uber.org/zap"
+
 	"github.com/rs/cors"
 	"github.com/valkey-io/valkey-go"
-	"go.uber.org/zap"
-	"gopkg.in/yaml.v3"
-
-	"github.com/yanolja/ogem"
+	"github.com/yanolja/ogem/config"
 	"github.com/yanolja/ogem/server"
 	"github.com/yanolja/ogem/state"
 	"github.com/yanolja/ogem/utils"
 	"github.com/yanolja/ogem/utils/env"
 )
-
-func loadConfig(path string, logger *zap.SugaredLogger) (*server.Config, error) {
-	// Setting default values
-	config := server.Config{
-		ValkeyEndpoint: "",
-		OgemApiKey:     "",
-		RetryInterval:  "1m",
-		PingInterval:   "1h",
-		Port:           8080,
-		Providers:      ogem.ProvidersStatus{},
-	}
-
-	// Checks if config is specified via environment variable.
-	configSource := env.OptionalStringVariable("CONFIG_SOURCE", path)
-	configToken := env.OptionalStringVariable("CONFIG_TOKEN", "")
-	configData, err := func(configSource string, configToken string) ([]byte, error) {
-		// Handle URL or local path
-		if strings.HasPrefix(configSource, "http://") || strings.HasPrefix(configSource, "https://") {
-			logger.Infow("Fetching remote config", "url", configSource)
-			return fetchRemoteConfig(configSource, configToken)
-		}
-		logger.Infow("Loading local config", "path", configSource)
-		return os.ReadFile(configSource)
-	}(configSource, configToken)
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to get config data: %v", err)
-	}
-
-	// Overrides config with the YAML data.
-	if err := yaml.Unmarshal(configData, &config); err != nil {
-		return nil, fmt.Errorf("failed to parse config: %v", err)
-	}
-
-	// Overrides config with environment variables.
-	// Therefore, the values from the environment variables precede the values from the YAML file.
-	config.ValkeyEndpoint = env.OptionalStringVariable("VALKEY_ENDPOINT", config.ValkeyEndpoint)
-	config.OgemApiKey = env.OptionalStringVariable("OPEN_GEMINI_API_KEY", config.OgemApiKey)
-	config.GenaiStudioApiKey = env.OptionalStringVariable("GENAI_STUDIO_API_KEY", config.GenaiStudioApiKey)
-	config.GoogleCloudProject = env.OptionalStringVariable("GOOGLE_CLOUD_PROJECT", config.GoogleCloudProject)
-	config.OpenAiApiKey = env.OptionalStringVariable("OPENAI_API_KEY", config.OpenAiApiKey)
-	config.ClaudeApiKey = env.OptionalStringVariable("CLAUDE_API_KEY", config.ClaudeApiKey)
-	config.RetryInterval = env.OptionalStringVariable("RETRY_INTERVAL", config.RetryInterval)
-	config.PingInterval = env.OptionalStringVariable("PING_INTERVAL", config.PingInterval)
-	config.Port = env.OptionalIntVariable("PORT", config.Port)
-
-	return &config, nil
-}
-
-func fetchRemoteConfig(url string, token string) ([]byte, error) {
-	client := &http.Client{
-		Timeout: 10 * time.Second,
-	}
-
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	if token != "" {
-		req.Header.Set("Authorization", "Bearer "+token)
-	}
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("failed to fetch config: HTTP %d", resp.StatusCode)
-	}
-
-	return io.ReadAll(resp.Body)
-}
-
-func setupStateManager(valkeyEndpoint string) (state.Manager, func(), error) {
-	if valkeyEndpoint == "" {
-		// Maximum memory usage of 2GB.
-		memoryManager, cleanup := state.NewMemoryManager(2 * 1024 * 1024 * 1024)
-		return memoryManager, cleanup, nil
-	}
-
-	valkeyClient, err := valkey.NewClient(valkey.ClientOption{
-		InitAddress: []string{valkeyEndpoint},
-	})
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to create Valkey client: %v", err)
-	}
-	return state.NewValkeyManager(valkeyClient), nil, nil
-}
 
 func main() {
 	logger := utils.Must(zap.NewProduction())
@@ -122,19 +29,19 @@ func main() {
 
 	configPath := flag.String("config", "config.yaml", "path to config file")
 	flag.Parse()
-	config, err := loadConfig(*configPath, sugar)
+	config, err := config.LoadConfig(*configPath, sugar)
 	if err != nil {
 		sugar.Fatalw("Failed to load config", "error", err)
 	}
 
-	stateManager, cleanup, err := setupStateManager(config.ValkeyEndpoint)
+	stateManager, cleanup, err := setupStateManager(config)
 	if err != nil {
 		sugar.Fatalw("Failed to setup state manager", "error", err)
 	}
 
 	sugar.Infow("Loaded config", "config", config)
 
-	proxy, err := server.NewProxyServer(stateManager, cleanup, *config, sugar)
+	proxy, err := server.NewProxyServer(stateManager, cleanup, config, sugar)
 	if err != nil {
 		sugar.Fatalw("Failed to create proxy server", "error", err)
 	}
@@ -142,20 +49,7 @@ func main() {
 	mux := http.NewServeMux()
 	mux.HandleFunc("/v1/chat/completions", proxy.HandleAuthentication(proxy.HandleChatCompletions))
 
-	corsMiddleware := cors.New(cors.Options{
-		AllowedOrigins: []string{"*"},
-		AllowedMethods: []string{"GET", "POST", "PUT", "DELETE", "OPTIONS"},
-		AllowedHeaders: []string{"*"},
-		Debug:          false,
-	})
-
-	port := env.OptionalStringVariable("PORT", "8080")
-	address := fmt.Sprintf(":%s", port)
-
-	httpServer := &http.Server{
-		Addr:    address,
-		Handler: corsMiddleware.Handler(mux),
-	}
+	httpServer := setupServer(config, mux)
 
 	shutdownSignal := make(chan os.Signal, 1)
 	signal.Notify(shutdownSignal, os.Interrupt, syscall.SIGTERM)
@@ -184,10 +78,43 @@ func main() {
 		}
 	}()
 
-	sugar.Infow("Starting server", "address", address)
+	sugar.Infow("Starting server", "address", httpServer.Addr)
 	if err := httpServer.ListenAndServe(); err != nil && err != http.ErrServerClosed {
 		sugar.Fatalw("Failed to start server", "error", err)
 	}
 
 	sugar.Infow("Server exited gracefully")
+}
+
+func setupStateManager(config *config.Config) (state.Manager, func(), error) {
+	if config.ValkeyEndpoint == "" {
+		// Maximum memory usage of 2GB.
+		memoryManager, cleanup := state.NewMemoryManager(2 * 1024 * 1024 * 1024)
+		return memoryManager, cleanup, nil
+	}
+
+	valkeyClient, err := valkey.NewClient(valkey.ClientOption{
+		InitAddress: []string{config.ValkeyEndpoint},
+	})
+	if err != nil {
+		return nil, nil, fmt.Errorf("failed to create Valkey client: %v", err)
+	}
+	return state.NewValkeyManager(valkeyClient), nil, nil
+}
+
+func setupServer(config *config.Config, handler http.Handler) *http.Server {
+	corsMiddleware := cors.New(cors.Options{
+		AllowedOrigins: []string{"*"},
+		AllowedMethods: []string{"GET", "POST", "PUT", "DELETE", "OPTIONS"},
+		AllowedHeaders: []string{"*"},
+		Debug:          false,
+	})
+
+	port := env.OptionalStringVariable("PORT", strconv.Itoa(config.Port))
+	address := fmt.Sprintf(":%s", port)
+
+	return &http.Server{
+		Addr:    address,
+		Handler: corsMiddleware.Handler(handler),
+	}
 }
